refactored standard greedy UFS
